requirement.txt
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
certifi==2025.1.31
comm==0.2.2
decorator==5.2.1
executing==2.2.0
goodfire==0.3.5
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
ipython==9.0.1
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
jedi==0.19.2
joblib==1.4.2
jupyterlab_widgets==3.0.13
matplotlib-inline==0.1.7
numpy==2.2.3
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.50
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.0.1
scikit-learn==1.6.1
scipy==1.15.2
sniffio==1.3.1
stack-data==0.6.3
threadpoolctl==3.5.0
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
widgetsnbextension==4.0.13
